Исправил текующие ошибки
