+++ js/main.js
@@ -1,4 +1,5 @@
-import {images} from './data.js';
+import {getImages} from './data.js';
 
+const images = getImages();
 // eslint-disable-next-line no-console
 console.log(images);
